allow MutableScripts to be directly executed
merges the shared logic into a base class to be inherited by Script and MutableScript; changed existing usages of Script to ExecutableScript.

This change allows you more flexibility in setting up your own Starscript compilation pipeline, and even the ability to precompile scripts and then just load their files (compile, save, stop the program, load the file, works as if it was just compiled) if you want; since all the Hypervisor expects is an RO span of Value constants and bytes for code, with a few helpers that exist on the base class.